+++ Tests/SwiftStdlibTests/IntExtensionsTests.swift
@@ -105,14 +105,6 @@ final class IntExtensionsTests: XCTestCase {
         XCTAssert(63.roundToNearest(25) == 75)
         XCTAssert(42.roundToNearest(0) == 42)
     }
-  
-    func testUpTo() {
-        var latestValue = 0
-        2.up(to: 5) { currentValue in
-            latestValue = currentValue
-        }
-        assert(latestValue == 5)
-    }
 
     func testOperators() {
         XCTAssertEqual(5 ** 2, 25)